refactor(store): separate into different files
